+++ lib/consistent_random/sidekiq_middleware.rb
@@ -8,6 +8,26 @@ class ConsistentRandom
       include Sidekiq::ServerMiddleware
     end
 
+    class << self
+      def install
+        Sidekiq.configure_server do |config|
+          config.server_middleware do |chain|
+            chain.prepend ConsistentRandom::SidekiqMiddleware
+          end
+
+          config.client_middleware do |chain|
+            chain.add ConsistentRandom::SidekiqClientMiddleware
+          end
+        end
+
+        Sidekiq.configure_client do |config|
+          config.client_middleware do |chain|
+            chain.add ConsistentRandom::SidekiqClientMiddleware
+          end
+        end
+      end
+    end
+
     def call(job_instance, job_payload, queue)
       ConsistentRandom.scope(job_payload["consistent_random_seed"]) do
         yield